Extract file identity interop partial
